backend/requirements.txt
requests
python-multipart
fastapi
uvicorn
python-jose
passlib[bcrypt]

